### 2019-06-08 #1498 @kthyng
No problem! I’ll work on it during the week.

On Sat, Jun 8, 2019 at 12:05 PM Lorena A. Barba <notifications@github.com>
wrote:

> @kthyng <https://github.com/kthyng> — Based on topic and current editor
> loads, I'd like to assign this submission to you. Can you take it? Let me
> know!
>
