### 2019-06-13 #1498 @kthyng
@tristan-salles is your paper missing a figure? In looking through, looks like a figure reference (left) but no figure.

also, do you have any [reviewer](https://bit.ly/joss-reviewers) recommendations? A handful of suggestions will be helpful so that I can ultimately find 2.
