### 2019-06-13 #1498 @kthyng
@rbeucher and @molgor Are you interested and available to review this submission to JOSS? We prefer reviews to be in within about 3 weeks and you can find guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
