### 2019-06-25 #1498 @kthyng
@hugoledoux Are you interested and available to review this submission to JOSS? We prefer reviews to be in within about 3 weeks and you can find guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
