### 2019-06-27 #1530 @whedon
[REVIEW]: bioLEC: A Python package to measure Landscape Elevational Connectivity
**Submitting author:** @tristan-salles (<a href="http://orcid.org/0000-0001-6095-7689">Tristan Salles</a>)
**Repository:** <a href="https://github.com/Geodels/bioLEC" target ="_blank">https://github.com/Geodels/bioLEC</a>
**Version:** 1.0.0
