### 2019-07-03 Geodels/bioLEC#1 @kbarnhart
JOSS Review: No GitHub version listed
The version described in the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/1530) is 1.0.0

There are presently no tags or releases listed on GitHub.

My assumption is that @tristan-salles is intentionally waiting to make changes indicated in review before tagging.
