### 2019-07-03 123516a
fix: scipy listed twice
