### 2019-07-03 178003a
fix: these test instructions worked for me.
