### 2019-07-03 64cbb09
docs: add an example of success message.
