### 2019-07-03 73e4ae6
fix: remove duplicate scipy from requirements.txt
