### 2019-07-03 Geodels/bioLEC#2 @kbarnhart
JOSS Review: Install instructions in many places
As part of my [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1530) I've looked through the `readme.md`, the read the docs page, the source code, and the notebooks. I've found that the installation instructions are duplicated across these locations.

I can understand wanting to make the installation instructions easily accessible. My experience, however, is that it is much better to have the instructions in one place and provide links. That way, when installation instructions change, you don't need to change all the files.

@tristan-salles feel free to make modifications based on this recommendation or completely ignore it.
