### 2019-07-03 Geodels/bioLEC#3 @kbarnhart
JOSS Review: Developer installation and test instructions almost worked
As part of my [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1530) I'm testing the installation instructions. As I'll note in my review in that repo I'm testing the developer install (but not the other install instructions).

I found a few small errors in the install and testing. A forthcoming PR will describe the minor fixes that worked for me.
