### 2019-07-03 Geodels/bioLEC#5 @kbarnhart
JOSS Review: Notebook file structure confusing
Another issue as part of my [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1530)

I cloned the repo and tried to run the notebooks locally.

After running `jupyter notebook` from the top level `Notebooks` directory I opened the `0-StartHere.ipynb` notebooks. I then ran into 404 errors with the links to the other notebooks. I assume this is because the are actually located in `bioLEC/src/bioLEC/Notebooks`

I'd recommend that there is one place for the file structure where the notebooks are located. If for some reason related to Docker that means they need to be in both of these places, perhaps make a note on the ReadTheDocs regarding how to run the notebooks locally).

I poked around in the repo trying to figure out a fix and found the `documentation.py` file. This may or not matter, but it was not at all clear to me whether the file is vestigial or whether I should have been able to use it to solve my 404 error issues.
