### 2019-07-03 Geodels/bioLEC#6 @kbarnhart
JOSS Review: Improvements to docstrings related to connectivity/boundary conditions
Comments from my [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1530) regarding the contents of the documentation.

@tristan-salles Overall I found the docstrings (and their representation on RTFD) really nice. There was, however, one case where I couldn't figure something out from the docs that I think I should have been able to figure out.

I was not able to figure out the periodic and symmetric keyword arguments without looking at the figure presented in the notebook `bioLEC1-genericCase.ipynb`. I would have liked to understand the values from the docstring alone, or to have an image reference to the notebook.

These two keywords brought up a few thoughts:
1) I think they are incompatible with one another, right? If so, the code presently doesn't raise an error if both are set to true.
2) (Perhaps not for this present release, but) I think it might make sense to consider each boundary (North, South, East, and West) independently. That is, you could have East and West boundaries as periodic, North as symmetric, and South as "normal".

In addition, I think it would be worth using a different name for the keyword argument `connected`. As I understand it, this pertains to D4/D8 connectivity. Perhaps "connectivity" (with valid values "axial" and "both") or "diagonals" with valid values (False, True). At the least, I'd recommend using the words D4 and D8 in the docstring and making the y direction paneling label in the `bioLEC1-genericCase.ipynb` figure reflect the keyword argument.
