### 2019-07-03 Geodels/bioLEC#7 @kbarnhart
JOSS Review: lavavu might not be working on Binder
As part of my [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1530) I'm testing the notebooks.

I couldn't get the parts that related to lavavu to run locally---but the rest of the notebooks worked for me just fine.

On Binder I got unexpected results for the cells that relate to lavavu (for example, the following screenshot).
![Screen Shot 2019-07-03 at 12 18 47 PM](https://user-images.githubusercontent.com/6589339/60617338-c5bf2280-9d90-11e9-9543-e6cf0567f6ed.png)

I tried this on both Safari and Chrome with the same failed results.
