### 2019-07-03 Geodels/bioLEC#8 @kbarnhart
JOSS Review: Links to API in RTFD
Another comment related to my [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1530)...

There are a number of places in the docs where I expected a link to the api documentation. For example the `landscapeConnectivity()` at the top of the [Quickstart](https://biolec.readthedocs.io/en/latest/usage.html#input).

The package is overall really well linked, this sort of issue was the only one where I wanted links that weren't there. I find that this sort of link helps a lot when one is trying to teach themself from the documentation.
