### 2019-07-03 Geodels/bioLEC#9 @kbarnhart
JOSS Review: Statement of need in RTFD
@tristan-salles I thought that the statement of need was very clear in the the `paper.md` but not as clear on the RTFD page.

[link to JOSS Review](https://github.com/openjournals/joss-reviews/issues/1530)
