### 2019-07-03 #1530 @kbarnhart
Reviewer update:

I've done a full pass at a review and was very pleased with the package. I think that bioLEC is a great contribution to topographic analysis and an appropriate contribution to JOSS.

I've made all of my comments as Issues. The three unchecked boxes in my checklist all correspond to items identified in issues. I will check them off as they are addressed.

I have only tested the developer and pip installs. I will not be able to test the Docker install or run the code on an HPCC. I was not 100% successful at running the notebooks locally or on Binder (both cases due to issues with `lavavu` indicated in the Issues made in the main repository.
