### 2019-07-07 554fb01
Remove installation instructions from the source code and notebook
