### 2019-07-07 Geodels/bioLEC#2 @tristan-salles
I think you're right @kbarnhart and I have decided to remove the installation instructions from the source code and the notebooks.

I kept it in the `readme.md` and the `readthedoc` pages as some people might only want to look quickly in the GitHub page for these instructions...
