### 2019-07-07 6f05735
Put examples/notebooks in a single place
