### 2019-07-07 cd51302
update path to images
