### 2019-07-08 cfb0bfd
add references
