### 2019-07-08 Geodels/bioLEC#5 @tristan-salles
I decided to duplicate the notebooks (they are actually really small in size less than 50kb in total) to ensure that the examples were available regardless off the use: local install/kitematic/binder.
