### 2019-07-08 Geodels/bioLEC#9 @tristan-salles
The statement of need has been updated in the `readthedoc` [page](https://biolec.readthedocs.io/en/latest/) to reflect the `paper.md`section.
