### 2019-07-08 88f6f16
Add references to API in the documentation
