### 2019-07-08 Geodels/bioLEC#8 @tristan-salles
I have added references to the API for the 3 main functions:
+ `landscapeConnectivity()`
+ `computeLEC()`
+ `writeLEC()`

in the [Quick start guide](https://biolec.readthedocs.io/en/latest/usage.html) documentation.
