### 2019-07-08 0631e3c
Update usage.rst
