### 2019-07-08 Geodels/bioLEC#6 @tristan-salles
As pointed by @kbarnhart, the _periodic_ and _symmetric_ boundary conditions are incompatible as they are applied to all boundaries in this current version of **bioLEC**. I have added a couple of lines in the code to raise a warning when both conditions are set to `True`. I have also added a link in the docstring to a figure in the documentation that illustrates both boundary conditions.

Also following @kbarnhart comments I have modified the keyword argument `connected` and changed it to `diagonals` with boolean values. I have also added the D4 and D8 words in the docstring plus I have modified the figure to reflect those changes.
