### 2019-07-08 e59d5a2
Improvements to docstrings related to connectivity/boundary conditions
