+++ Notebooks/notebooks/runLEC.py
@@ -18,7 +18,7 @@ parser.add_argument('-p','--periodic',help='True/false option for periodic bound
 parser.add_argument('-s','--symmetric',help='True/false option for symmetric boundary conditions', required=False, action="store_true", default=False)
 parser.add_argument('-w','--width',help='Float option for species niche width percentage', required=False, action="store_true", default=0.1)
 parser.add_argument('-f','--fix',help='Float option for species niche width fix values', required=False, action="store_true", default=None)
-parser.add_argument('-c','--connected',help='True/false option for computing the path based on the diagonal moves as well as the axial ones', required=False, action="store_true", default=True)
+parser.add_argument('-c','--diagonals',help='True/false option for computing the path based on the diagonal moves as well as the axial ones eg. D4/D8 connectivity', required=False, action="store_true", default=True)
 parser.add_argument('-n','--nout',help='Number for output frequency during run', required=False, action="store_true", default=500)
 parser.add_argument('-d','--delimiter',help='String for elevation grid csv delimiter', required=False,action="store_true",default=' ')
 parser.add_argument('-l','--level',help='Float for sea level position', required=False,action="store_true",default=-1.e6)
@@ -34,13 +34,13 @@ if args.verbose and rank == 0:
   print("   + Symmetric boundary conditions for the elevation grid: {}".format(args.symmetric))
   print("   + Species niche width percentage based on elevation extent: {}".format(args.width))
   print("   + Species niche width based on elevation extent: {}".format(args.fix))
-  print("   + Computes the path based on the diagonal moves as well as the axial ones: {}".format(args.connected))
+  print("   + Computes the path based on the diagonal moves as well as the axial ones: {}".format(args.diagonals))
   print("   + Elevation grid csv delimiter: {}".format(args.delimiter))
   print("   + Sea level position: {}".format(args.level))
   print("   + Number for output frequency: {}\n".format(args.nout))
 
 biodiv = LEC.landscapeConnectivity(filename=args.input,periodic=args.periodic,symmetric=args.symmetric,
-                                    sigmap=args.width,sigmav=args.fix,connected=args.connected,
+                                    sigmap=args.width,sigmav=args.fix,diagonals=args.diagnonals,
                                     delimiter=args.delimiter,sl=args.level)
 
 biodiv.computeLEC(args.nout)
