### 2019-07-08 #1530 @tristan-salles
Thanks, @kbarnhart for the really useful review, I have been updating the code and documentation to account for the comments and improvements that were proposed.
