+++ src/bioLEC/documentation.py
@@ -24,22 +24,12 @@ def install_documentation(path="./bioLEC-Notebooks"):
     and will overwrite any existing files with which they collide. The default
     path ("./bioLEC-Notebooks") is chosen to make collision less likely / problematic
     The documentation for bioLEC is in the form of jupyter notebooks.
-    Some dependencies exist for the notebooks to be useful:
-       - matplotlib: for some of the diagrams
-       - lavavu: interactive viewer for an alternative means of viewing cartesian_meshes
-       - cartopy: for plotting map examples
-    bioLEC dependencies are explicitly imported into the notebooks including:
-       - numpy
-       - scipy (for k-d tree point location)
-       - skimage
-       - pandas
-       - mpi4py
     """
 
     ## Question - overwrite or not ? shutils fails if directory exists.
 
     Notebooks_Path = _pkg_resources.resource_filename('bioLEC', 'Notebooks')
-    
+
     ct = _dir_util.copy_tree(Notebooks_Path, path,preserve_mode=1, preserve_times=1, preserve_symlinks=1, update=0, verbose=1, dry_run=0)
 
     return
