### 2019-07-08 Geodels/bioLEC#7 @tristan-salles
`lavavu` was working in `binder` but there is a continual battle between _jupyter hub security_ and the various reverse-proxy strategies required to ensure `lavavu` service visible.

As `lavavu` isn't required to run the **bioLEC** package I have decided to remove it from the `binder` user documentation but to keep it in the `Docker container`...
