### 2019-07-08 #1530 @kthyng
thanks @kbarnhart!

@hugoledoux Just a friendly reminder about your JOSS review.
