### 2019-07-09 #1530 @kbarnhart
@tristan-salles glad the review was helpful.

@kthyng I've completed my review and am happy to indicate that I think the contribution should be published in JOSS. I have left only one checkbox un-checked (related to versioning). As @tristan-salles indicated in [bioLEC Issue 1](https://github.com/Geodels/bioLEC/issues/1) the version will be tagged at the end of the JOSS review process.
