### 2019-07-15 #1530 @kthyng
@hugoledoux Want to check in again. If you know when you can work on this review, I will wait to ping you again until after that time. Thanks!
