### 2019-07-18 #1530 @hugoledoux
I clicked above what is working, and left some blanks, these are linked to the comments below.

## version
the github page doesn't have an official release (nor tags!).

## installation
the first install returned an error because I didn't have numpy installed, I needed to install it myself. You should install it as a dependencies (other packages were installed automatically). I see now it's in the `requirements.txt` file...hmmmm... Python packaging is messy, not sure why this happened.

## unit tests
Is it all in file `/src/tests/testInstall.py`? This works and all tests are fine, but is it enough?

## examples?
Perhaps putting a folder with a few terrains and one example file would be a good idea?
