+++ src/setup.py
@@ -18,12 +18,17 @@ if __name__ == "__main__":
           long_description_content_type='text/markdown',
           packages          = ['bioLEC'],
           install_requires  = [
-                        'numpy',
-                        'scipy>=0.15.0',
-                        'pandas',
-                        'mpi4py',
+                        'numpy>=1.15.0',
+                        'six>=1.11.0',
+                        'setuptools>=38.4.0',
+                        'mpi4py>=3.0.0',
+                        'pandas>=0.24',
+                        'scipy>=1.2',
+                        'scikit-image>=0.15',
+                        'pyevtk>=1.1.1',
+                        'matplotlib>=3.0',
                         'rasterio>=1.0.23',
-                        'scikit-image'],
+                        'nose>=1.3'],
           python_requires   = '>=2.7, >=3.5',
           package_data      = {'bioLEC': ['Notebooks/notebooks/*ipynb',
                                           'Notebooks/notebooks/*py',
