+++ src/requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.15
+numpy>=1.15.0
 six>=1.11.0
 setuptools>=38.4.0
 mpi4py>=3.0.0
