### 2019-07-19 #1530 @tristan-salles
Thanks for the review @hugoledoux !!!

## version

I was waiting for the end of the JOSS review process before creating the **bioLEC** version. This is now done and its tagged as [ **v1.0.0**](https://github.com/Geodels/bioLEC/releases/tag/v1.0.0) to match the Python package version.

## installation

Hum.. this is weird, I did try a clean install and it worked fine but I have seen a few posts on google suggesting to put the complete version number such as `numpy>=1.15.0` instead of only `numpy>=1.15` and to make sure that the version number between `requirements.txt` and `setup.py` were matching... I have modified both files accordingly, just in case it clears the dependency issue you've experienced...

## unit tests

the `testInstall.py` is where the unit test is done. I agree this is a simple test but it actually:
1. returns any library installation problem,
2. runs all the functions via the `test flag` in the `LEC.py` file and
3. checks the obtained results for a given topography against the expected ones.

## examples?

First it is worth mentioning that the third jupyter notebook `bioLEC3-MeshGen` explains how to generate terrains from standard ARCGIS like format and `bioLEC1-genericCase` shows how to use 2D numpy arrays as input for `bioLEC`. But I agree with @hugoledoux, additional examples are a good idea and I have added 2 additional terrains to test for the user :
+ `dataset/pyrenees.csv`
+ `dataset/afghan.csv`

I have updated the `bioLEC2-Kosciuszko` jupyter notebook to add the link to these new topographic files. I think we need to make sure the package size is not too big in size and I believe having 3 different terrains might be enough.
