### 2019-07-23 #1530 @hugoledoux
All is good, I've clicked them all above 👍🏽

Two small caveats:
1. I wouldn't use "ArcGIS formats", mostly because that's a for-profit company and this is not accurate (unless you support only ArcGIS...). Your third jupyter book speaks of "GeoTIFF", which I much prefer personally!
2. The main link to the mybinder.org only has links to only 2 notebooks, but one can see it in the `notebooks`.
