### 2019-07-23 #1530 @tristan-salles
Thanks, I have updated the binder notebook links
