### 2019-07-23 #1530 @kthyng
Ok @tristan-salles now's the time to make a Zenodo archive, and report the DOI in the review thread.
