### 2019-07-24 #1530 @tristan-salles
sorry didn't check it... I've updated it now.

[![DOI](https://zenodo.org/badge/185154268.svg)](https://zenodo.org/badge/latestdoi/185154268)
