### 2019-07-24 #1530 @kthyng
@openjournals/joss-eics This paper is ready to be taken over! Should be all ready for you!

congrats @tristan-salles and thanks @kbarnhart and @hugoledoux!!
