### 2019-07-25 e4f41cf
Update paper.bib
Fixes to references
