+++ joss/paper.bib
@@ -21,7 +21,7 @@
 }
 
 @article{Clubb:2017,
-   author = {Clubb, F.J. and Mudd, S. M. and  Milodowski, D. T. and Valters, D. A. and  Slater, L. J. and Hurst, M.D.  and Limaye, A. B.},
+   author = {Clubb, F.J. and Mudd, S. M. and  Milodowski, D. T. and Valters, D. A. and  Slater, L. J. and Hurst, M. D.  and Limaye, A. B.},
    title = {Geomorphometric delineation of floodplains and terraces from objectively defined topographic thresholds},
    journal = {Earth Surface Dynamics},
    year = 2017,
@@ -31,7 +31,7 @@
 }
 
 @article{Elsen:2015,
-   author = {{Elsen}, P.~R. and {Tingley}, M.~W.},
+   author = {Elsen, P. R. and Tingley, M. W.},
    title = {Global mountain topography and the fate of montane species under climate change},
    journal = {Nature Climate Change},
    year = 2015,
@@ -72,7 +72,7 @@
 }
 
 @article{Salles:2018,
-  title = {pyBadlands: A framework to simulate sediment transport, landscape dynamics and basin stratigraphic evolution through space and time},
+  title = {{pyBadlands}: A framework to simulate sediment transport, landscape dynamics and basin stratigraphic evolution through space and time},
   author = {Salles, Tristan and Ding, Xuesong and Brocard, Gilles},
   journal = {PLOS ONE},
   year = {2018},
@@ -83,8 +83,8 @@
 }
 
 @article{vanderWalt:2014,
-  title = {Scikit Image Contributors - scikit-image: image processing in Python},
-  author = {van der Walt, S. and Schonberger, J.L. and  Nunez-Iglesias, J. and Boulogne, F. and Warner, J.D. and Yager, N. and Gouillart, E. and Yu, T.},
+  title = {scikit-image: image processing in {Python}},
+  author = {{van der Walt}, S. and Schonberger, J. L. and  Nunez-Iglesias, J. and Boulogne, F. and Warner, J. D. and Yager, N. and Gouillart, E. and Yu, T. and {the scikit image contributors}},
   journal = {PeerJ},
   year = {2014},
   volume = {2:e453},
