### 2019-07-25 #1530 @kyleniemeyer
Hi @tristan-salles, please merge the PRs I just submitted that fix a few things in the paper: https://github.com/Geodels/bioLEC/pull/11 and https://github.com/Geodels/bioLEC/pull/10, then we'll be good to go.
