### 2019-07-25 #1530 @kyleniemeyer
Congrats on your submission's publication in JOSS @tristan-salles!

Many thanks to @kbarnhart and @hugoledoux for reviewing, and @kthyng for editing!
